backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.76",
    "datasets>=4.8.4",
    "fastapi>=0.135.2",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.13",
    "notebook>=7.5.5",
    "pgvector>=0.4.2",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pycocotools>=2.0.11",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.37",
    "modal>=0.73.0",
    "uvicorn>=0.42.0",
    "python-multipart>=0.0.22",
    "langchain-anthropic>=1.4.0",
]
